Builder/MyIoTProject/requirements.txt
alembic>=1.0.6
cffi>=1.11.5
Click>=7.0
defusedxml>=0.5.0
Flask>=1.0.2
Flask-Login>=0.4.1
Flask-Migrate>=2.3.1
Flask-Restless>=0.17.0
Flask-SocketIO>=3.1.2
Flask-SQLAlchemy>=2.3.2
Flask-WTF>=0.14.2
flask-restless-swagger-2>=0.0.3
GeoAlchemy2>=0.5.0
itsdangerous>=1.1.0
Jinja2>=2.10
lxml>=4.3.0
Mako>=1.0.7
MarkupSafe>=1.1.0
mimerender>=0.6.0
ordered-set>=3.1
paho-mqtt>=1.4.0
psycopg2>=2.7.7
pycparser>=2.19
pyecore>=0.9.0
python-dateutil>=2.7.5
python-editor>=1.0.3
python-engineio>=3.2.3
python-mimeparse>=1.6.0
python-socketio>=3.1.1
RestrictedPython>=4.0b7
six>=1.12.0
SQLAlchemy>=1.2.16
Werkzeug>=0.14.1
WTForms>=2.2.1
Flask-Script
